+++ CHANGELOG.md
@@ -8,3 +8,16 @@
 - Added dataset: MELD
 
 ## 2026-03-31
+
+* Added NOTES.MD
+* Added folder: optimization with apso.py
+* Added folder: architectures
+* Created config file
+* Created vision_net.py with the CNN architecture
+* Added folder: data_pipeline
+* Created loaders.py with a loader for FER2013
+* Deleted notebooks folder
+* models folder
+* plots folder
+* Obtained best parameters for vision model
+* Obtained best vision model